move email-sent into auth-form
