+++ src/routes/auth/(auth-form)/email-sent.svelte
@@ -11,11 +11,11 @@
   const props: Props = $props();
 </script>
 
-<section class="w-[min(400px,100%)] rounded-lg p-4 border-1">
-  <section class="mb-5">
+<section class="card w-[min(400px,100%)]">
+  <header>
     <h2>{props.title}</h2>
     <p class="mt-1">{props.description}</p>
-  </section>
+  </header>
 
   <section class="flex flex-col gap-1">
     <p>{props.action}</p>
